requirements.txt
# MoE-DiReF 依赖
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
PyYAML>=6.0
tqdm>=4.65.0

# 阶段 8 评估（可选）：LPIPS 与 FID 更准确
# pip install lpips
